Fix mismatched tax class for configurable products
When calculating taxes for configurable products, the tax class of the simple product was used during calculations but the tax class of the parent was used when reporting transactions.  This PR standardizes tax classes to always use the tax class assigned to the simple product first, and if no tax class is set to fall back to the parent.