Add product tax code product attribute (#153)
* Add product tax code product attribute

Adds a new product attribute called "tj_ptc" that is used to store a product tax code.  When set, this PTC is sent to the API.  When undefined, the tax class is used as a fallback.

* Add a PTC attribute source model

Add a PTC attribute source model that can be used in a dropdown select, replacing the text input.

* Add default category to dropdown, fix child ptc lookup

Adds a default "Fully Taxable" option with an empty value to the top of the dropdown.  Bugfix for sending an invalid order id.  Correctly lookup the child PTC for configurable products when syncing tranx.