Handling adjustment fees for $0 refunds (#166)
* Fix syncing $0 refunds

When distributing refund adjustment fees over the line items, we divide by the subtotal.  In the case where the subtotal is exactly 0, this causes an error.  This PR skips distributing adjustment fees over line items when the subtotal is 0.

Note: If an adjustment fee is defined, and the subtotal is $0, the adjustment fee will not be included.  However manually including the adjustment fee as a line item results in an incorrect expected tax amount.